Upgrades Fix - Web Adapter Configuration Fails if IWA is enabled + Federation Resource Failures + Data store test resource failures fix + Wrong Server URL endpoints displayed at the end

+++ Modules/ArcGIS/ArcGIS.psm1
@@ -318,7 +318,7 @@ function Get-ArcGISURL
     {
         $ServerWAMachineNode = ($AllNodes | Where-Object { ($_.Role -icontains 'ServerWebAdaptor')} | Select-Object -First 1)
         $ServerWAMachineName = $ServerWAMachineNode.NodeName
-        $ServerHostName = Get-FQDN $ServerWAMachineName
+        $ServerExternalDNSName = Get-FQDN $ServerWAMachineName
         if(($ServerWAMachineNode.SslCertifcates | Where-Object { $_.Target -icontains 'WebAdaptor'}  | Measure-Object).Count -gt 0)
         {
             $SSLCertificate = $ServerWAMachineNode.SslCertifcates | Where-Object { $_.Target -icontains 'WebAdaptor' }  | Select-Object -First 1
@@ -712,21 +712,38 @@ function Configure-ArcGIS
                         }
 
                         if($JobFlag){
-                            Write-Host "Dot Sourcing the Configuration:- ArcGISFederation"
-                            . "$PSScriptRoot\Configuration\ArcGISFederation.ps1" -Verbose:$false
+
+                            $RemoteFederation = if($ConfigurationParamsHashtable.ConfigData.Federation){$true}else{$false}
+            
+                            $PortalServerFederation = $False
                             
-                            Write-Host "Compiling the Configuration:- ArcGISFederation"
-                            ArcGISFederation -ConfigurationData $ConfigurationParamsHashtable 
-                            
-                            if($Credential){
-                                $JobFlag = Start-DSCJob -ConfigurationName ArcGISFederation -Credential $Credential -DebugMode $DebugMode
-                            }else{
-                                
-                                $JobFlag = Start-DSCJob -ConfigurationName ArcGISFederation -DebugMode $DebugMode
+                            if(-not($RemoteFederation))
+                            {
+                                $ServerCheck = (($ConfigurationParamsHashtable.AllNodes | Where-Object { $_.Role -icontains 'Server' }  | Measure-Object).Count -gt 0)
+                                $PortalCheck = (($ConfigurationParamsHashtable.AllNodes | Where-Object { $_.Role -icontains 'Portal' }  | Measure-Object).Count -gt 0)
+                                if($ServerCheck -and $PortalCheck)
+                                {
+                                    $PortalServerFederation = $True
+                                }
                             }
 
-                            if(Test-Path ".\ArcGISFederation") {
-                                Remove-Item ".\ArcGISFederation" -Force -ErrorAction Ignore -Recurse
+                            if($RemoteFederation -or $PortalServerFederation){
+                                Write-Host "Dot Sourcing the Configuration:- ArcGISFederation"
+                                . "$PSScriptRoot\Configuration\ArcGISFederation.ps1" -Verbose:$false
+                                
+                                Write-Host "Compiling the Configuration:- ArcGISFederation"
+                                ArcGISFederation -ConfigurationData $ConfigurationParamsHashtable 
+                                
+                                if($Credential){
+                                    $JobFlag = Start-DSCJob -ConfigurationName ArcGISFederation -Credential $Credential -DebugMode $DebugMode
+                                }else{
+                                    
+                                    $JobFlag = Start-DSCJob -ConfigurationName ArcGISFederation -DebugMode $DebugMode
+                                }
+
+                                if(Test-Path ".\ArcGISFederation") {
+                                    Remove-Item ".\ArcGISFederation" -Force -ErrorAction Ignore -Recurse
+                                }
                             }
                         }
